refactor: Extract download annotation logic to function
